backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chroma-hnswlib>=0.7.6",
    "chromadb>=1.0.13",
    "crewai>=0.130.0",
    "fastapi>=0.115.13",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.25",
    "langchain-groq>=0.3.2",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "selenium>=4.33.0",
    "unstructured>=0.17.2",
    "watchdog>=6.0.0",
]
